Prevent infinite loop when loggly endpoint is not reachable
When loggly is not reachable in cases where loggly will be
blocked by browser extensions for example, in some clumsy
constellations it's possible that infinite loops can occur.
In order to prevent infinite loops, the library is configured
by a defined size of parallel request slots and suppression
of further requests.

Original Author: Micha